chat_browser/components/parsers/parser_type_b.py

from .base_parser import BaseParser
from bs4 import BeautifulSoup


class ParserTypeB(BaseParser):
    def parse(self, html_content: str) -> dict:
        soup = BeautifulSoup(html_content, 'html.parser')
        message_div = soup.find('div', {'data-message-author-role': True})
        if not message_div:
            return {}  # Not suitable for this parser

        message = {
            "message": {
                "author_role": message_div['data-message-author-role'],
                "id": message_div['data-message-id'],
                "model_slug": message_div.get('data-message-model-slug', ''),
                "content": self.extract_content(message_div)
            }
        }
        return message

    def extract_content(self, message_div):
        # Implement extraction logic based on the HTML structure
        content = {
            "type": "code_list",
            "items": []  # Populate based on parsed content
        }
        # Example extraction logic
        return content
