chat_browser/requirements.txt
streamlit
PyYAML
beautifulsoup4
